app/requirements.txt
beautifulsoup4==4.9.3
pymongo==3.11.4
python-dateutil==2.8.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.4
